+++ src/JSEngine.cpp
@@ -30,12 +30,20 @@ namespace be {
         return JS_NewStringLen(ctx, buff, size) ;
     }
     static JSValue js_ArrayBuffer_toArray(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) {
+        bool signed_int = false ;
+        if(argc>0) {
+            signed_int = JS_ToBool(ctx, argv[0]) ;
+        }
         size_t size ;
         char * buff = (char *)JS_GetArrayBuffer(ctx, &size, this_val) ;
         JSValue array = JS_NewArray(ctx) ;
         if(buff) {
             for(int i=0;i<(int)size;i++) {
-                JS_SetPropertyUint32(ctx, array, i, JS_NewUint32(ctx, buff[i])) ;
+                if(signed_int) {
+                    JS_SetPropertyUint32(ctx, array, i, JS_NewInt32(ctx, (int8_t)buff[i])) ;
+                } else {
+                    JS_SetPropertyUint32(ctx, array, i, JS_NewUint32(ctx, buff[i])) ;
+                }
             }
         }
         return array ;