+++ changelog.md
@@ -1,5 +1,17 @@
 ## dev
 
+Added:
+
+* uart.unsetup()
+* uart.isInstalled()
+* sensor driver TMP117
+
+Fixed:
+
+Refactored:
+
+* uart.write() 可以接受数组参数
+* ArrayBuffer.prototype.toArray() 增加参数 signed
 
 ## v0.3.4   2024-6-4